Move RootLieftimeScope settings to VContainerSettings
