--- VContainer/Assets/VContainer/Editor/CodeGen.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b1fbc4ba07964adc8ecfa036ac830d42
-timeCreated: 1594288504
