--- VContainer/Assets/VContainer/Editor/VContainerSettings.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cf05ae6ecc0242b681226a809b30c3ae
-timeCreated: 1595484573
