+++ VContainer/Assets/VContainer/Editor/VContainerSettingsEditor.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 086d1f02a80d4633aaced41d23477067
-timeCreated: 1595492413
+fileFormatVersion: 2
+guid: a71622b13e9fb4ff28937885bb67dcc4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
