+++ VContainer/Assets/VContainer/Runtime/Unity/VContainerSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9cf4d8df1d704d5689f3c69bd70d1cb9
+timeCreated: 1595759041
